move `FakesUserGroups` into `Statamic\Testing\Concerns`
`setTestUserGroups()` becomes `protected` to match the other concerns.
